requirements.txt
Django==3.2.20
paho-mqtt==1.6.1
dj-database-url==0.5.0
kombu==5.2.3
python-json-logger==2.0.4
gunicorn==20.1.0
requests==2.31.0
psycopg2-binary==2.9.6
celery==5.2.7
redis==4.3.4
flower==1.0.0
SQLAlchemy==1.4.46
django-timezone-field==5.0
python-crontab==2.7.1
python-dateutil==2.8.2
amqp==5.1.1
